Use Float Instead Of Int For Rating

+++ app/src/main/java/pl/edu/pja/s27773/filmoteka/model/dto/MovieDto.kt
@@ -11,6 +11,6 @@ data class MovieDto (
     val releaseDate: LocalDate,
     val category: Category,
     val status: Status,
-    val rating: Int?, // nullable for NOT_WATCHED
+    val rating: Float?, // nullable for NOT_WATCHED
     val posterUri: Uri?
 )
